modified few routes

+++ routes/adminRoutes.js
@@ -13,7 +13,7 @@ adminRoute.get('/users', verifyToken, isAdmin, async (req, res) => {
     try {
         const users = await User.find({});
         if (!users) {
-            return res.status(400).json({ message: "Database is empty" });
+            return res.status(404).json({ message: "Database is empty" });
         }
         res.status(200).json({ users: users });
     } catch (error) {
@@ -25,7 +25,7 @@ adminRoute.get('/books', verifyToken, isAdmin, async (req, res) => {
     try {
         const users = await Book.find({});
         if (!users) {
-            return  res.status(400).json({ message: "Database is empty" });
+            return  res.status(404).json({ message: "Database is empty" });
         }
         res.status(200).json({ books: users });
     } catch (error) {
@@ -37,7 +37,7 @@ adminRoute.get('/authors', verifyToken, isAdmin, async (req, res) => {
     try {
         const users = await Author.find({});
         if (!users) {
-            return  res.status(400).json({ message: "Database is empty" });
+            return  res.status(404).json({ message: "Database is empty" });
         }
         res.status(200).json({ authors: users });
     } catch (error) {
@@ -49,7 +49,7 @@ adminRoute.post('/add_user', verifyToken, isAdmin, async (req, res) => {
     try {
         const email = await User.findOne({ 'email': req.body.email }, 'email');
         if (email) {
-            return res.status(400).json({ message: "User with this e-mail exists in database" });
+            return res.status(404).json({ message: "User with this e-mail exists in database" });
         }
         else{
             const pass = bcrypt.hashSync(req.body.password, salt);
@@ -71,7 +71,7 @@ adminRoute.post('/add_book', verifyToken, isAdmin, async (req, res) => {
     try {
         const isbn = await Book.findOne({'isbn': req.body.isbn});
         if(isbn){
-            return res.status(400).json({"message" : "Book with this isbn exists in database"});
+            return res.status(404).json({message : "Book with this isbn exists in database"});
         }
         const newBook = new Book({
             title: req.body.title,
@@ -98,7 +98,7 @@ adminRoute.post('/add_author', verifyToken, isAdmin, async (req, res) => {
         const author = await Author.findOne({'fullName': req.body.fullName,
          'email': req.body.email});
         if(author){
-            return res.status(400).json({"message": "Author with this data exists in database"});
+            return res.status(400).json({message: "Author with this data exists in database"});
         }
         const newAuthor = new Author({
             fullName: req.body.fullName,
@@ -120,7 +120,7 @@ adminRoute.put('/add_admin/:id', verifyToken, isAdmin, async (req, res) => {
         const userId = req.params.id;
         const user = await User.findById(userId);
         if (!user) {
-            return  res.status(400).json({ message: "USer not found" });
+            return  res.status(404).json({ message: "User not found" });
         }
         user.isAdmin = true;
         await user.save();
@@ -135,7 +135,7 @@ adminRoute.put('/remove_admin/:id', verifyToken, isAdmin, async (req, res) => {
     try {
         const user = await User.findById(req.params.id);
         if (!user) {
-           return res.status(400).json({ message: "User not found" });
+           return res.status(404).json({ message: "User not found" });
         }
         if (!user.isAdmin) {
             return res.status(400).json({ message: "This user is not admin" });
@@ -157,6 +157,9 @@ adminRoute.put('/update_book/:id', verifyToken, isAdmin, async (req, res) => {
             return res.status(400).json({ message: "Invalid id_ for this book" });
         }
         const book = await Book.findById(id);
+        if(!book){
+            return res.status(404).json({message: "Book not found"});
+        }
         book.title = req.body.title || book.title;
         book.isbn = req.body.isbn || book.isbn;
         book.pages = req.body.pages || book.pages;
@@ -180,8 +183,11 @@ adminRoute.put('/update_author/:id', verifyToken, isAdmin, async (req, res) => {
         if (!mongoose.Types.ObjectId.isValid(id)) {
             return res.status(400).json({ message: "Invalid id_ for this author" });
         }
-        const author = await Author.findById(id);
 
+        const author = await Author.findById(id);
+        if(!author){
+            return res.status(404).json({message: "Author not found"});
+        }
         author.fullName = req.body.fullName;
         author.age = req.body.age;
         author.email = req.body.email;
@@ -198,7 +204,7 @@ adminRoute.put("/change_password/:id", verifyToken, isAdmin, async (req, res) =>
     try {
         const user = await User.findById(req.params.id);
         if (!user) {
-            return  res.status(400).json({ message: "User not found" });
+            return  res.status(404).json({ message: "User not found" });
         }
         const newPass = bcrypt.hashSync(req.body.password, salt);
 
@@ -218,7 +224,7 @@ adminRoute.put('/update_user/:userId', verifyToken, isAdmin, async (req, res) =>
         const user = await User.findById(id);
 
         if (!user) {
-            return res.status(400).json({ message: "User not found" });
+            return res.status(404).json({ message: "User not found" });
         }
 
         user.username = req.body.username || user.username;
@@ -267,9 +273,13 @@ adminRoute.delete('/delete_author/:id', verifyToken, isAdmin, async (req, res) =
 adminRoute.delete('/delete_user/:userId', verifyToken, async (req, res) => {
     try {
         const id = req.params.userId;
+        const userIsAdmin = await User.findById(id);
+        if(userIsAdmin.isAdmin){
+            return res.status(400).json({message:"This user is admin"});
+        }
         const user = await User.findByIdAndDelete(id);
         if (!user) {
-            return  res.status(400).json({ message: "User not found" });
+            return  res.status(404).json({ message: "User not found" });
         }
 
         res.status(200).json({ message: "User successfully deleted" });
